+++ pyscriptjs/src/main.ts
@@ -14,7 +14,6 @@ import { createCustomElements } from './components/elements';
 import { UserError, ErrorCode, _createAlertBanner } from './exceptions';
 import { type Stdio, StdioMultiplexer, DEFAULT_STDIO } from './stdio';
 import { PyTerminalPlugin } from './plugins/pyterminal';
-import { PyListPlugin } from './plugins/pylist';
 import { SplashscreenPlugin } from './plugins/splashscreen';
 import { ImportmapPlugin } from './plugins/importmap';
 import { StdioDirector as StdioDirector } from './plugins/stdiodirector';
@@ -88,7 +87,6 @@ export class PyScriptApp {
             new SplashscreenPlugin(),
             new PyTerminalPlugin(this),
             new ImportmapPlugin(),
-            new PyListPlugin(),
         );
 
         this._stdioMultiplexer = new StdioMultiplexer();
