+++ pyscriptjs/src/python/pyscript/_html.py
@@ -1,5 +1,6 @@
 import time
 from textwrap import dedent
+from datetime import datetime as dt
 
 import js
 
@@ -279,3 +280,21 @@ class PyListTemplate:
     def child_appended(self, child):
         """Overwrite me to define logic"""
         pass
+
+
+class PyItem(PyItemTemplate):
+    def on_click(self, evt=None):
+        self.data["done"] = not self.data["done"]
+        self.strike(self.data["done"])
+
+        self.select("input").element.checked = self.data["done"]
+
+
+class PyList(PyListTemplate):
+    item_class = PyItem
+
+    def add(self, item):
+        if isinstance(item, str):
+            item = {"content": item, "done": False, "created_at": dt.now()}
+
+        super().add(item, labels=["content"], state_key="done")
